refactor(ui): extract formatting and architecture logic from workspace page
Extract review-workspace-formatting.ts from page.tsx (1,076 → 950 lines):
- Formatting helpers (formatCodeRegion, formatAnalysisDuration, etc.)
- Architecture column computation (buildArchitectureColumns)
- Architecture constants and types

Closes #173
